@@ -8,7 +8,7 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.1",
+        "php": "^8.1",
         "bavix/laravel-wallet": "^5.3",
         "fideloper/proxy": "^4.2",
         "fruitcake/laravel-cors": "^2.0",
@@ -16,7 +16,7 @@
         "irazasyed/telegram-bot-sdk": "^3.4.1",
         "laravel-notification-channels/telegram": "^0.5.0",
         "laravel/framework": "^8.0",
-        "laravel/horizon": "^5.9",
+        "laravel/horizon": "^v5.9.3",
         "laravel/socialite": "^5.0",
         "laravel/tinker": "^v2.7.1",
         "laravel/ui": "^3.4.5",
