@@ -2,15 +2,11 @@
 
 namespace App\Providers;
 
-use App\Notifications\ConfirmOrder;
-use Illuminate\Auth\Events\Registered;
 use Illuminate\Foundation\Support\Providers\EventServiceProvider as ServiceProvider;
-use Illuminate\Support\Facades\Event;
 use App\Events\OrderAccepted;
 use App\Events\OrderConfirmed;
-use App\Listeners\SendAcceptedNotifications;
+use App\Listeners\AcceptedNotificationsListener;
 use App\Listeners\ConfirmOrderListener;
-use App\Listeners\SendConfirmedNotifications;
 
 class EventServiceProvider extends ServiceProvider
 {
@@ -23,13 +19,6 @@ class EventServiceProvider extends ServiceProvider
         \SocialiteProviders\Manager\SocialiteWasCalled::class => [
             // ... other providers
             'SocialiteProviders\\Telegram\\TelegramExtendSocialite@handle',
-        ],
-        OrderAccepted::class => [
-            SendAcceptedNotifications::class,
-        ],
-        OrderConfirmed::class => [
-            ConfirmOrderListener::class,
-            SendConfirmedNotifications::class
         ],
     ];
 
