@@ -40,6 +40,6 @@
         <p style="color:#347AF0">Ожидание перевода</p>
     </div>
     <div class="text-block">
-        <p>Как только пользователь подтвердит отправку средств, вам поступит уведомление в телеграме с последними 4 цифрами номера/счета, с которого были отправлены средства.</p>
+        <p>Как только пользователь подтвердит отправку средств, вам поступит уведомление в телеграме.</p>
     </div>
 </div>
