@@ -41,7 +41,7 @@
     </div>
 
     <div class="text-block">
-        <p>Пользователь подтвердил перевод средств. Последние 4 цифры номера карты (счета) отправителя *{{$order->comment}}. Как только средства поступят, подтвердите получение.</p>
+        <p>Пользователь подтвердил перевод средств. Как только средства поступят, подтвердите получение.</p>
     </div>
     <div class="footer">
         <a class="button button-blue" onclick="confirmOrder()">Подтвердить получение</a>
