@@ -86,7 +86,7 @@
     </div>
     <h1>Ошибка 404</h1>
     <p>Извините! Страница, которую Вы ищете, не может быть найдена</p>
-    <a href="{{env('APP_URL')}}" class="btn">Перейти в кошелек</a>
+    <a href="/wallet/" class="btn">Перейти в кошелек</a>
 </div>
 </body>
 </html>
