@@ -38,6 +38,7 @@
                         <a onclick="decline('{{$order->id}}');" style="margin-top:40px;" class="button button-danger">Отменить заявку</a>
                         @elseif ($order->status == 'accepted')
                             <div class="created-block form-inline">
+                                @if ($order->currency != 'TON')
                                 <p>Отправьте <strong><span class="step2-amount">{{$order->amount}} </span> <span class="step2-currency"> {{$order->currency}}</span> </strong>в <strong>{{$order->payment}}</strong>@if (App\Models\Payment::where('title', $order->payment)->firstOrFail()->crypto) на адрес:@else по номеру карты:@endif</p>
                                 <div class="w-100">
                                     <div class="col-12">
@@ -69,11 +70,49 @@
                                         </div>
                                     </div>
                                 </div>
+                                @else
+                                <p>Заявка #{{$order->id}} принята кипером. Переведите {{$order->amount}} TON по следующим реквизитам:</p>
+                                    <p class="w-100 mt-3 d-md-flex align-items-center">Адрес: <a class="wallet-link copy-link copy-link-small" data-toggle="popover" data-placement="bottom" data-content="Ссылка скопирована в буфер обмена." data-original-title="" title="">
+                                            <span>{{$order->payment_details}}</span>
+                                            <svg width="33" height="36" viewBox="0 0 33 36" fill="none" xmlns="http://www.w3.org/2000/svg">
+                                                <g opacity="0.9">
+                                                    <path d="M30.5379 5.84501H29.9899C27.6071 5.84501 25.6755 3.91338 25.6755 1.53059C25.6755 0.685295 24.9902 0 24.1449 0H1.53059C0.685295 0 0 0.685295 0 1.53059V28.6242C0 29.4695 0.685295 30.1548 1.53059 30.1548H2.07854C4.46127 30.1548 6.39285 32.0864 6.39285 34.4691C6.39285 35.3144 7.07815 35.9997 7.92344 35.9997H30.5379C31.3832 35.9997 32.0685 35.3144 32.0685 34.4691V7.37559C32.0685 6.5303 31.3832 5.84501 30.5379 5.84501ZM4.72707 27.0936C3.80702 27.0936 3.06117 26.3478 3.06117 25.4277V8.06117C3.06117 5.29975 5.29975 3.06117 8.06117 3.06117H21.2223C21.991 3.06117 22.6142 3.68436 22.6142 4.45309C22.6142 5.22182 21.991 5.84501 21.2223 5.84501H7.92354C7.07825 5.84501 6.39296 6.5303 6.39296 7.37559V25.4277C6.39296 26.3478 5.64711 27.0936 4.72707 27.0936ZM29.0073 27.9386C29.0073 30.7001 26.7687 32.9386 24.0073 32.9386H14.4541C11.6927 32.9386 9.45413 30.7001 9.45413 27.9386V13.9062C9.45413 11.1448 11.6927 8.90618 14.4541 8.90618H24.0073C26.7687 8.90618 29.0073 11.1448 29.0073 13.9062V27.9386Z" fill="url(#paint1_linear)"></path>
+                                                </g>
+                                                <defs>
+                                                    <linearGradient id="paint1_linear" x1="16.0342" y1="-16.1999" x2="33.7999" y2="-11.6723" gradientUnits="userSpaceOnUse">
+                                                        <stop stop-color="#FF9134"></stop>
+                                                        <stop offset="1" stop-color="#E72269"></stop>
+                                                    </linearGradient>
+                                                </defs>
+                                            </svg>
+                                        </a></p>
+                                    <p class="w-100 mt-3 d-md-flex align-items-center">Примечание (мемо): <a class="wallet-link copy-link copy-link-small" data-toggle="popover" data-placement="bottom" data-content="Ссылка скопирована в буфер обмена." data-original-title="" title="">
+                                            <span>{{$order->comment}}</span>
+                                            <svg width="33" height="36" viewBox="0 0 33 36" fill="none" xmlns="http://www.w3.org/2000/svg">
+                                                <g opacity="0.9">
+                                                    <path d="M30.5379 5.84501H29.9899C27.6071 5.84501 25.6755 3.91338 25.6755 1.53059C25.6755 0.685295 24.9902 0 24.1449 0H1.53059C0.685295 0 0 0.685295 0 1.53059V28.6242C0 29.4695 0.685295 30.1548 1.53059 30.1548H2.07854C4.46127 30.1548 6.39285 32.0864 6.39285 34.4691C6.39285 35.3144 7.07815 35.9997 7.92344 35.9997H30.5379C31.3832 35.9997 32.0685 35.3144 32.0685 34.4691V7.37559C32.0685 6.5303 31.3832 5.84501 30.5379 5.84501ZM4.72707 27.0936C3.80702 27.0936 3.06117 26.3478 3.06117 25.4277V8.06117C3.06117 5.29975 5.29975 3.06117 8.06117 3.06117H21.2223C21.991 3.06117 22.6142 3.68436 22.6142 4.45309C22.6142 5.22182 21.991 5.84501 21.2223 5.84501H7.92354C7.07825 5.84501 6.39296 6.5303 6.39296 7.37559V25.4277C6.39296 26.3478 5.64711 27.0936 4.72707 27.0936ZM29.0073 27.9386C29.0073 30.7001 26.7687 32.9386 24.0073 32.9386H14.4541C11.6927 32.9386 9.45413 30.7001 9.45413 27.9386V13.9062C9.45413 11.1448 11.6927 8.90618 14.4541 8.90618H24.0073C26.7687 8.90618 29.0073 11.1448 29.0073 13.9062V27.9386Z" fill="url(#paint1_linear)"></path>
+                                                </g>
+                                                <defs>
+                                                    <linearGradient id="paint1_linear" x1="16.0342" y1="-16.1999" x2="33.7999" y2="-11.6723" gradientUnits="userSpaceOnUse">
+                                                        <stop stop-color="#FF9134"></stop>
+                                                        <stop offset="1" stop-color="#E72269"></stop>
+                                                    </linearGradient>
+                                                </defs>
+                                            </svg>
+                                        </a></p>
+                                    <p class="mt-3">При отправке средств укажите примечание (мемо).<br /><br />
+                                        <b>ДЕПОЗИТ НЕ БУДЕТ ЗАЧИСЛЕН БЕЗ ПРИМЕЧАНИЯ! </b></p>
+                                    <p class="mt-3">Как только средства будут зачислены, вы получите уведомление. Среднее время зачисления средств – одна минута.</p>
+
+                                @endif
                                 <div class="d-flex accept_buttons w-100 justify-content-between" style="margin-top:40px;">
-                                    <a onclick="acceptSending('{{$order->id}}');" class="button button-blue button-accept">Переведено, далее...</a>
+                                    @if ($order->currency == 'TON')
+                                        <a href="{{ "ton://transfer/" . $order->payment_details . "?amount=" . ($order->amount * 1000000000) . "&text=" . $order->comment}}" class="button button-blue">Открыть Toncoin Кошелек</a>
+                                    @else
+                                        <a onclick="acceptSending('{{$order->id}}');" class="button button-blue button-accept">Переведено, далее...</a>
+                                    @endif
                                     <a onclick="decline('{{$order->id}}');" style="" class="button button-danger">Отменить заявку</a>
                                 </div>
-                                <p class="mt-3">* для подтверждения отправки средств приготовьте последние 4 цифры счета / номера карты</p>
 
                             </div>
                             @elseif ($order->status == 'pending')
@@ -157,25 +196,21 @@
 
         function acceptSending(id) {
             let _token = $('meta[name="csrf-token"]').attr('content');
-            let comment = prompt('Чтобы подтвердить отправление средств, ведите последние 4 цифры номера карты или счета, с которого были отправлены средства')
-            if (comment) {
-                $.ajax({
-                    url: "/api/orders/acceptSendingByUser",
-                    type:"POST",
-                    data:{
-                        _token: _token,
-                        id: id,
-                        comment: comment
-                    },
-                    error: function () {
-                        alert('При попытке подтвердить заявку произошла ошибка')
-                    },
-                    success:function(response){
-                        alert('Вы успешно подтвердили отправку средств')
-                        window.location.href = '/dashboard/orders/' + id
-                    },
-                });
-            }
+            $.ajax({
+                url: "/api/orders/acceptSendingByUser",
+                type:"POST",
+                data:{
+                    _token: _token,
+                    id: id,
+                },
+                error: function () {
+                    alert('При попытке подтвердить заявку произошла ошибка')
+                },
+                success:function(response){
+                    alert('Вы успешно подтвердили отправку средств')
+                    window.location.href = '/dashboard/orders/' + id
+                },
+            });
         }
 
         $(document).ready(function(){
