@@ -41,25 +41,25 @@ Route::group(['prefix' => 'wallet'], function () {
 
 Route::group(['prefix' => 'dashboard'], function () {
     Route::get('/', [App\Http\Controllers\WalletController::class, 'index'])->name('dashboard');
-    Route::get('/profile', [App\Http\Controllers\WalletController::class, 'profile'])->name('wallet.profile');
-    Route::get('/explorer', [App\Http\Controllers\WalletController::class, 'explorer'])->name('wallet.explorer');
-    Route::get('/orders/{id}', [App\Http\Controllers\WalletController::class, 'order'])->name('wallet.pages.order');
-    Route::get('/transfer', [App\Http\Controllers\WalletController::class, 'transfer'])->name('wallet.transfer');
+    Route::get('/profile', [App\Http\Controllers\WalletController::class, 'profile'])->name('dashboard.profile');
+    Route::get('/explorer', [App\Http\Controllers\WalletController::class, 'explorer'])->name('dashboard.explorer');
+    Route::get('/orders/{id}', [App\Http\Controllers\WalletController::class, 'order'])->name('dashboard.pages.order');
+    Route::get('/transfer', [App\Http\Controllers\WalletController::class, 'transfer'])->name('dashboard.transfer');
 
     // ADMIN PAGES
     Route::post('/currencies/{slug}', [App\Http\Controllers\WalletController::class, 'updateCurrency'])->middleware('admin');
-    Route::get('/orders/', [App\Http\Controllers\WalletController::class, 'orders'])->middleware('gate_manager')->name('wallet.orders');
-    Route::get('/stages/', [App\Http\Controllers\WalletController::class, 'stages'])->middleware('admin')->name('wallet.stages');
-    Route::get('/reports/', [App\Http\Controllers\WalletController::class, 'reports'])->middleware('admin')->name('wallet.reports');
-    Route::get('/hft/', [App\Http\Controllers\WalletController::class, 'hft'])->middleware('admin')->name('wallet.hft');
-    Route::get('/currencies/', [App\Http\Controllers\WalletController::class, 'currencies'])->middleware('admin')->name('wallet.currencies');
-    Route::get('/currencies/{slug}', [App\Http\Controllers\WalletController::class, 'currency'])->middleware('admin')->name('wallet.currency');
-    Route::get('/payments/', [App\Http\Controllers\WalletController::class, 'payments'])->middleware('admin')->name('wallet.payments');
-    Route::get('/users', [App\Http\Controllers\WalletController::class, 'getUsers'])->name('wallet.users')->middleware('gate_manager');
-    Route::get('/gates', [App\Http\Controllers\WalletController::class, 'gates'])->name('wallet.gates')->middleware('gate_manager');
-    Route::get('/tags', [App\Http\Controllers\WalletController::class, 'tags'])->name('wallet.tags')->middleware('admin');
-    Route::get('/telegram', [App\Http\Controllers\WalletController::class, 'telegram'])->name('wallet.telegram')->middleware('admin');
-    Route::get('/settings', [App\Http\Controllers\WalletController::class, 'settings'])->name('wallet.settings')->middleware('admin');
+    Route::get('/orders/', [App\Http\Controllers\WalletController::class, 'orders'])->middleware('gate_manager')->name('dashboard.orders');
+    Route::get('/stages/', [App\Http\Controllers\WalletController::class, 'stages'])->middleware('admin')->name('dashboard.stages');
+    Route::get('/reports/', [App\Http\Controllers\WalletController::class, 'reports'])->middleware('admin')->name('dashboard.reports');
+    Route::get('/hft/', [App\Http\Controllers\WalletController::class, 'hft'])->middleware('admin')->name('dashboard.hft');
+    Route::get('/currencies/', [App\Http\Controllers\WalletController::class, 'currencies'])->middleware('admin')->name('dashboard.currencies');
+    Route::get('/currencies/{slug}', [App\Http\Controllers\WalletController::class, 'currency'])->middleware('admin')->name('dashboard.currency');
+    Route::get('/payments/', [App\Http\Controllers\WalletController::class, 'payments'])->middleware('admin')->name('dashboard.payments');
+    Route::get('/users', [App\Http\Controllers\WalletController::class, 'getUsers'])->name('dashboard.users')->middleware('gate_manager');
+    Route::get('/gates', [App\Http\Controllers\WalletController::class, 'gates'])->name('dashboard.gates')->middleware('gate_manager');
+    Route::get('/tags', [App\Http\Controllers\WalletController::class, 'tags'])->name('dashboard.tags')->middleware('admin');
+    Route::get('/telegram', [App\Http\Controllers\WalletController::class, 'telegram'])->name('dashboard.telegram')->middleware('admin');
+    Route::get('/settings', [App\Http\Controllers\WalletController::class, 'settings'])->name('dashboard.settings')->middleware('admin');
 });
 
 
@@ -71,7 +71,6 @@ Route::post('/api/tags/add', [App\Http\Controllers\SystemApiController::class, '
 Route::post('/api/tags/remove', [App\Http\Controllers\SystemApiController::class, 'removeTag'])->middleware('admin');
 Route::post('/api/start_token_sale/', [App\Http\Controllers\SystemApiController::class, 'startTokenSale'])->middleware('admin');
 Route::post('/api/generate_user_wallets/', [App\Http\Controllers\SystemApiController::class, 'generateUserWallets'])->middleware('admin');
-Route::post('/api/orders/{id}/confirm', [App\Http\Controllers\SystemApiController::class, 'confirmOrder'])->middleware('admin');
 Route::post('/api/orders/{id}/decline', [App\Http\Controllers\SystemApiController::class, 'declineOrder'])->middleware('gate_manager');
 Route::post('/api/withdraw-payment', [App\Http\Controllers\SystemApiController::class, 'withdrawPayment'])->middleware('admin');
 Route::post('/api/send', [App\Http\Controllers\SystemApiController::class, 'sendTokens'])->middleware('admin');
